refactor: consolidated InvalidAddressError into @taquito/core
re #1993
