+++ packages/taquito-contracts-library/src/errors.ts
@@ -1,12 +1,4 @@
-/**
- *  @category Error
- *  @description Error that indicates an invalid address being used or passed
- */
-export class InvalidAddressError extends Error {
-  constructor(message: string) {
-    super(message);
-  }
-}
+export { InvalidAddressError } from '@taquito/core';
 
 /**
  *  @category Error
