+++ packages/taquito-contracts-library/test/taquito-contracts-library.spec.ts
@@ -1,8 +1,9 @@
-import { InvalidAddressError, InvalidScriptFormatError } from '../src/errors';
+import { InvalidScriptFormatError } from '../src/errors';
 import { ContractsLibrary } from '../src/taquito-contracts-library';
 import { entrypoints, entrypoints2 } from './data/contract-entrypoints';
 import { script, script2 } from './data/contract-script';
 import { VERSION } from '../src/version';
+import { InvalidAddressError } from '@taquito/core';
 
 describe('ContractsLibrary tests', () => {
   it('ContractsLibrary is instantiable', () => {
@@ -106,7 +107,7 @@ describe('ContractsLibrary tests', () => {
       })
     ).toThrow(
       expect.objectContaining({
-        message: expect.stringContaining('Address is invalid: KTinvalid'),
+        message: expect.stringContaining(`The address 'KTinvalid' is invalid.`),
       })
     );
   });
