+++ packages/taquito-rpc/package.json
@@ -67,6 +67,7 @@
     ]
   },
   "dependencies": {
+    "@taquito/core": "^16.1.2-beta-RC.0",
     "@taquito/http-utils": "^16.1.2-beta-RC.0",
     "@taquito/utils": "^16.1.2-beta-RC.0",
     "bignumber.js": "^9.1.0"
