+++ packages/taquito-rpc/src/rpc-client-modules/rpc-cache.ts
@@ -52,9 +52,8 @@ import {
   PendingOperations,
   OriginationProofParams,
 } from '../types';
-
+import { InvalidAddressError } from '@taquito/core';
 import {
-  InvalidAddressError,
   InvalidContractAddressError,
   validateContractAddress,
   validateAddress,
@@ -76,7 +75,6 @@ type RpcMethodParam =
   | PendingOperationsQueryArguments
   | EndorsingRightsQueryArguments
   | OriginationProofParams;
-
 
 const defaultTtl = 1000;
 
@@ -1237,22 +1235,22 @@ export class RpcClientCache implements RpcClientInterface {
    * @default args { version: '1', applied: true, refused: true, outdated, true, branchRefused: true, branchDelayed: true, validationPass: undefined }
    * @see https://tezos.gitlab.io/CHANGES.html?highlight=pending_operations#id4
    */
-    async getPendingOperations(
-      args: PendingOperationsQueryArguments = {}
-    ): Promise<PendingOperations> {
-      const key = this.formatCacheKey(
-        this.rpcClient.getRpcUrl(),
-        RPCMethodName.GET_PENDING_OPERATIONS,
-        [args]
-      );
-      if (this.has(key)) {
-        return this.get(key);
-      } else {
-        const response = this.rpcClient.getPendingOperations(args);
-        this.put(key, response);
-        return response;
-      }
+  async getPendingOperations(
+    args: PendingOperationsQueryArguments = {}
+  ): Promise<PendingOperations> {
+    const key = this.formatCacheKey(
+      this.rpcClient.getRpcUrl(),
+      RPCMethodName.GET_PENDING_OPERATIONS,
+      [args]
+    );
+    if (this.has(key)) {
+      return this.get(key);
+    } else {
+      const response = this.rpcClient.getPendingOperations(args);
+      this.put(key, response);
+      return response;
     }
+  }
 
   /**
    *
