+++ packages/taquito-rpc/src/taquito-rpc.ts
@@ -64,13 +64,8 @@ import {
   OriginationProofParams,
 } from './types';
 import { castToBigNumber } from './utils/utils';
-import {
-  InvalidAddressError,
-  validateAddress,
-  validateContractAddress,
-  ValidationResult,
-} from '@taquito/utils';
-
+import { validateAddress, validateContractAddress, ValidationResult } from '@taquito/utils';
+import { InvalidAddressError } from '@taquito/core';
 export { castToBigNumber } from './utils/utils';
 
 export {
