+++ packages/taquito-sapling/package.json
@@ -69,6 +69,7 @@
     "@airgap/sapling-wasm": "0.0.9",
     "@stablelib/nacl": "^1.0.3",
     "@stablelib/random": "^1.0.1",
+    "@taquito/core": "^16.1.2-beta-RC.0",
     "@taquito/rpc": "^16.1.2-beta-RC.0",
     "@taquito/taquito": "^16.1.2-beta-RC.0",
     "@taquito/utils": "^16.1.2-beta-RC.0",
