+++ packages/taquito-utils/package.json
@@ -63,6 +63,7 @@
     ]
   },
   "dependencies": {
+    "@taquito/core": "^16.1.2-beta-RC.0",
     "@stablelib/blake2b": "^1.0.1",
     "@stablelib/ed25519": "^1.0.3",
     "@types/bs58check": "^2.1.0",
