+++ packages/taquito-utils/src/errors.ts
@@ -1,3 +1,5 @@
+export { InvalidAddressError } from '@taquito/core';
+
 /**
  *  @category Error
  *  @description Error that indicates an invalid key being passed or used
@@ -58,19 +60,6 @@ export class InvalidContractAddressError extends Error {
   public name = 'InvalidContractAddressError';
   constructor(public contractAddress: string) {
     super(`The contract address '${contractAddress}' is invalid`);
-  }
-}
-
-/**
- *  @category Error
- *  @description Error that indicates an invalid address being passed or used (both contract and implicit)
- */
-export class InvalidAddressError extends Error {
-  public name = 'InvalidAddressError';
-  constructor(public address: string, errorDetail?: string) {
-    super();
-    const baseMessage = `The address '${address}' is invalid.`;
-    this.message = errorDetail ? `${baseMessage} ${errorDetail}` : baseMessage;
   }
 }
