+++ packages/taquito/package.json
@@ -77,6 +77,7 @@
     ]
   },
   "dependencies": {
+    "@taquito/core": "^16.1.2-beta-RC.0",
     "@taquito/http-utils": "^16.1.2-beta-RC.0",
     "@taquito/local-forging": "^16.1.2-beta-RC.0",
     "@taquito/michel-codec": "^16.1.2-beta-RC.0",
