+++ packages/taquito/src/tz/rpc-tz-provider.ts
@@ -9,9 +9,9 @@ import {
   validateAddress,
   ValidationResult,
   validateKeyHash,
-  InvalidAddressError,
   InvalidKeyHashError,
 } from '@taquito/utils';
+import { InvalidAddressError } from '@taquito/core';
 
 export class RpcTzProvider extends OperationEmitter implements TzProvider {
   constructor(context: Context) {
