+++ packages/taquito/test/estimate/rpc-estimate-provider.spec.ts
@@ -28,7 +28,7 @@ import {
 } from '../contract/helper';
 import { OpKind, PvmKind } from '@taquito/rpc';
 import { TransferTicketParams } from '../../src/operations/types';
-import { InvalidAddressError } from '@taquito/utils';
+import { InvalidAddressError } from '@taquito/core';
 import { ContractAbstraction } from '../../src/contract';
 import { genericMultisig } from '../../../../integration-tests/data/multisig';
 import { RpcContractProvider } from '../../src/contract/rpc-contract-provider';
